Change resources from passive accrual to one-time capture bonus
Reverses M2's ticking mechanic per feedback: territory resource
numbers are a lump sum credited to the capturing player's wallet
once, immediately - not a per-minute rate that grows over time.
ResourceStockpile drops lastSettledAt/derive-on-read entirely; it's
now just a plain wallet incremented by creditTerritoryResources(),
wired into both the initial random starting-territory assignment and
every subsequent claim.

Verified against live game data: claiming a territory credited the
exact resource amounts shown on that territory, and the wallet
stayed flat afterward (no further ticking). Existing players' wallets
were recomputed from their currently-owned territories so the switch
is consistent immediately rather than leaving stale accrued values.

+++ src/main/java/com/canarda/readiness/game/dto/ResourceStockpileResponse.java
@@ -4,9 +4,9 @@ import com.canarda.readiness.game.domain.ResourceType;
 
 import java.util.Map;
 
-public record ResourceStockpileResponse(Long playerId, Map<ResourceType, Double> amounts) {
+public record ResourceStockpileResponse(Long playerId, Map<ResourceType, Integer> amounts) {
 
-    public static ResourceStockpileResponse of(Long playerId, Map<ResourceType, Double> amounts) {
+    public static ResourceStockpileResponse of(Long playerId, Map<ResourceType, Integer> amounts) {
         return new ResourceStockpileResponse(playerId, amounts);
     }
 }
